Fixed a bug with incorrect Redis password data type and slightly modified database models.

+++ database/models/referal.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Integer, String
+from sqlalchemy import BigInteger, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,7 +7,7 @@ from .base import BaseModel
 class ReferalModel(BaseModel):
     __tablename__ = "referals"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     inviter_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     code: Mapped[str] = mapped_column(String(10), nullable=True, server_default="usr")
